#297 refactoring of algorithm class
