+++ sources/Algo/include/SlackNodeAlgorithm.h
@@ -1,0 +1,53 @@
+//
+// Copyright (c) 2022, RTE (http://www.rte-france.com)
+// See AUTHORS.txt
+// All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+//
+
+/**
+ * @file  SlackNodeAlgorithm.h
+ *
+ * @brief Dynaflow launcher algorithms for determination of slack header file
+ *
+ */
+
+#pragma once
+
+#include "NetworkManager.h"
+#include "Node.h"
+
+namespace dfl {
+
+using NodePtr = std::shared_ptr<inputs::Node>;  ///< Alias for pointer to node
+namespace algo {
+
+/**
+ * @brief Algorithm to perform on nodes to find the slack node
+ */
+class SlackNodeAlgorithm {
+ public:
+  /**
+   * @brief Constructor
+   *
+   * @param slackNode the slack node to update with the algorithm
+   */
+  explicit SlackNodeAlgorithm(NodePtr& slackNode);
+
+  /**
+  * @brief Perform elementary step to determine the slack node
+  *
+  * The used criteria: the slack node is the node which has the higher voltage level then the higher number of connected nodes (in that order)
+  *
+  * @param node the node to process
+  */
+  void operator()(const NodePtr& node);
+
+ private:
+  NodePtr& slackNode_;  ///< The slack node to update
+};
+}  // namespace algo
+}  // namespace dfl
