+++ sources/Algo/src/GeneratorDefinitionAlgorithm.cpp
@@ -1,0 +1,144 @@
+//
+// Copyright (c) 2022, RTE (http://www.rte-france.com)
+// See AUTHORS.txt
+// All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+//
+
+#include "GeneratorDefinitionAlgorithm.h"
+#include "Log.h"
+#include "Message.hpp"
+#include <DYNCommon.h>
+
+namespace dfl {
+namespace algo {
+
+GeneratorDefinitionAlgorithm::GeneratorDefinitionAlgorithm(Generators& gens, BusGenMap& busesWithDynamicModel,
+                                                           const inputs::NetworkManager::BusMapRegulating& busMap, bool infinitereactivelimits,
+                                                           const boost::shared_ptr<DYN::ServiceManagerInterface>& serviceManager) :
+    generators_(gens),
+    busesWithDynamicModel_(busesWithDynamicModel),
+    busMap_(busMap),
+    useInfiniteReactivelimits_{infinitereactivelimits},
+    serviceManager_(serviceManager) {}
+
+void
+GeneratorDefinitionAlgorithm::operator()(const NodePtr& node) {
+  auto& node_generators = node->generators;
+
+  auto isModelWithInvalidDiagram = [](GeneratorDefinition::ModelType model, inputs::Generator generator) {
+    return (model == GeneratorDefinition::ModelType::DIAGRAM_PQ_SIGNALN || model == GeneratorDefinition::ModelType::REMOTE_DIAGRAM_PQ_SIGNALN ||
+            model == GeneratorDefinition::ModelType::PROP_DIAGRAM_PQ_SIGNALN) &&
+           !isDiagramValid(generator);
+  };
+
+  for (const auto& generator : node_generators) {
+    auto it = busMap_.find(generator.regulatedBusId);
+    assert(it != busMap_.end());
+    auto nbOfRegulatingGenerators = it->second;
+    GeneratorDefinition::ModelType model = GeneratorDefinition::ModelType::SIGNALN;
+    if (node_generators.size() == 1 && IsOtherGeneratorConnectedBySwitches(node)) {
+      model = useInfiniteReactivelimits_ ? GeneratorDefinition::ModelType::PROP_SIGNALN : GeneratorDefinition::ModelType::PROP_DIAGRAM_PQ_SIGNALN;
+      if (!isModelWithInvalidDiagram(model, generator)) {
+        busesWithDynamicModel_.insert({generator.regulatedBusId, generator.id});
+      }
+    } else {
+      switch (nbOfRegulatingGenerators) {
+      case dfl::inputs::NetworkManager::NbOfRegulating::ONE:
+        if (generator.regulatedBusId == generator.connectedBusId) {
+          model = useInfiniteReactivelimits_ ? GeneratorDefinition::ModelType::SIGNALN : GeneratorDefinition::ModelType::DIAGRAM_PQ_SIGNALN;
+        } else {
+          model = useInfiniteReactivelimits_ ? GeneratorDefinition::ModelType::REMOTE_SIGNALN : GeneratorDefinition::ModelType::REMOTE_DIAGRAM_PQ_SIGNALN;
+        }
+        break;
+      case dfl::inputs::NetworkManager::NbOfRegulating::MULTIPLES:
+        model = useInfiniteReactivelimits_ ? GeneratorDefinition::ModelType::PROP_SIGNALN : GeneratorDefinition::ModelType::PROP_DIAGRAM_PQ_SIGNALN;
+        if (!isModelWithInvalidDiagram(model, generator)) {
+          busesWithDynamicModel_.insert({generator.regulatedBusId, generator.id});
+        }
+        break;
+      default:  //  impossible by definition of the enum
+        break;
+      }
+    }
+    if (isModelWithInvalidDiagram(model, generator)) {
+      continue;
+    }
+    generators_.emplace_back(generator.id, model, node->id, generator.points, generator.qmin, generator.qmax, generator.pmin, generator.pmax, generator.targetP,
+                             generator.regulatedBusId);
+  }
+}
+
+bool
+GeneratorDefinitionAlgorithm::isDiagramValid(const inputs::Generator& generator) {
+  // If there are no points, the diagram will be constructed from the pmin, pmax, qmin and qmax values.
+  // We check the validity of pmin,pmax and qmin,qmax values
+  if (generator.points.empty()) {
+    if (DYN::doubleEquals(generator.pmin, generator.pmax)) {
+      LOG(warn) << MESS(InvalidDiagramAllPEqual, generator.id) << LOG_ENDL;
+      return false;
+    }
+    if (DYN::doubleEquals(generator.qmin, generator.qmax)) {
+      LOG(warn) << MESS(InvalidDiagramQminsEqualQmaxs, generator.id) << LOG_ENDL;
+      return false;
+    }
+    return true;
+  }
+
+  // If there is only one point, the diagram is not valid
+  if (generator.points.size() == 1) {
+    LOG(warn) << MESS(InvalidDiagramOnePoint, generator.id) << LOG_ENDL;
+    return false;
+  }
+
+  auto firstP = generator.points.front().p;
+  bool allQminEqualQmax = true;
+  bool allPEqual = true;
+  auto it = generator.points.begin();
+  while ((allQminEqualQmax || allPEqual) && it != generator.points.end()) {
+    allQminEqualQmax = allQminEqualQmax && it->qmin == it->qmax;
+    allPEqual = allPEqual && it->p == firstP;
+    ++it;
+  }
+  bool valid = !allQminEqualQmax && !allPEqual;
+
+  if (!valid) {
+    if (allQminEqualQmax && allPEqual) {
+      LOG(warn) << MESS(InvalidDiagramBothError, generator.id) << LOG_ENDL;
+    } else if (allQminEqualQmax) {
+      LOG(warn) << MESS(InvalidDiagramQminsEqualQmaxs, generator.id) << LOG_ENDL;
+    } else if (allPEqual) {
+      LOG(warn) << MESS(InvalidDiagramAllPEqual, generator.id) << LOG_ENDL;
+    }
+  }
+  return valid;
+}
+
+bool
+GeneratorDefinitionAlgorithm::IsOtherGeneratorConnectedBySwitches(const NodePtr& node) const {
+  auto vl = node->voltageLevel.lock();
+  auto buses = serviceManager_->getBusesConnectedBySwitch(node->id, vl->id);
+
+  if (buses.size() == 0) {
+    return false;
+  }
+
+  for (const auto& id : buses) {
+    auto found = std::find_if(vl->nodes.begin(), vl->nodes.end(), [&id](const NodePtr& node) { return node->id == id; });
+#ifdef _DEBUG_
+    // shouldn't happen by construction of the elements
+    assert(found != vl->nodes.end());
+#endif
+    if ((*found)->generators.size() != 0) {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+}  // namespace algo
+}  // namespace dfl
