+++ sources/Algo/src/MainConnexComponentAlgorithm.cpp
@@ -1,0 +1,47 @@
+//
+// Copyright (c) 2022, RTE (http://www.rte-france.com)
+// See AUTHORS.txt
+// All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+//
+
+#include "MainConnexComponentAlgorithm.h"
+
+namespace dfl {
+namespace algo {
+
+MainConnexComponentAlgorithm::MainConnexComponentAlgorithm(ConnexGroup& mainConnexity) : markedNodes_{}, mainConnexity_(mainConnexity) {}
+
+void
+MainConnexComponentAlgorithm::updateConnexGroup(ConnexGroup& group, const std::vector<NodePtr>& nodes) {
+  for (const auto& node : nodes) {
+    if (markedNodes_.count(node) == 0) {
+      markedNodes_.insert(node);
+      group.push_back(node);
+      updateConnexGroup(group, (node)->neighbours);
+    }
+  }
+}
+
+void
+MainConnexComponentAlgorithm::operator()(const NodePtr& node) {
+  if (markedNodes_.count(node) > 0) {
+    // already processed
+    return;
+  }
+
+  ConnexGroup group;
+  markedNodes_.insert(node);
+  group.push_back(node);
+  updateConnexGroup(group, node->neighbours);
+
+  if (mainConnexity_.size() < group.size()) {
+    mainConnexity_.swap(group);
+  }
+}
+
+}  // namespace algo
+}  // namespace dfl
