+++ sources/Outputs/include/Dyd.h
@@ -17,8 +17,16 @@
 
 #pragma once
 
-#include "Algo.h"
+#include "DynModelDefinitionAlgorithm.h"
 #include "DynamicDataBaseManager.h"
+#include "GeneratorDefinitionAlgorithm.h"
+#include "HVDCDefinitionAlgorithm.h"
+#include "LineDefinitionAlgorithm.h"
+#include "LoadDefinitionAlgorithm.h"
+#include "MainConnexComponentAlgorithm.h"
+#include "SVarCDefinitionAlgorithm.h"
+#include "ShuntDefinitionAlgorithm.h"
+#include "SlackNodeAlgorithm.h"
 
 #include <DYDBlackBoxModel.h>
 #include <DYDDynamicModelsCollection.h>
