+++ sources/Outputs/include/DydEvent.h
@@ -17,8 +17,17 @@
 
 #pragma once
 
-#include "Algo.h"
 #include "Contingencies.h"
+#include "ContingencyValidationAlgorithm.h"
+#include "DynModelDefinitionAlgorithm.h"
+#include "GeneratorDefinitionAlgorithm.h"
+#include "HVDCDefinitionAlgorithm.h"
+#include "LineDefinitionAlgorithm.h"
+#include "LoadDefinitionAlgorithm.h"
+#include "MainConnexComponentAlgorithm.h"
+#include "SVarCDefinitionAlgorithm.h"
+#include "ShuntDefinitionAlgorithm.h"
+#include "SlackNodeAlgorithm.h"
 
 #include <DYDBlackBoxModel.h>
 #include <DYDDynamicModelsCollection.h>
