+++ sources/Outputs/include/ParEvent.h
@@ -17,9 +17,18 @@
 
 #pragma once
 
-#include "Algo.h"
 #include "Configuration.h"
 #include "Contingencies.h"
+#include "ContingencyValidationAlgorithm.h"
+#include "DynModelDefinitionAlgorithm.h"
+#include "GeneratorDefinitionAlgorithm.h"
+#include "HVDCDefinitionAlgorithm.h"
+#include "LineDefinitionAlgorithm.h"
+#include "LoadDefinitionAlgorithm.h"
+#include "MainConnexComponentAlgorithm.h"
+#include "SVarCDefinitionAlgorithm.h"
+#include "ShuntDefinitionAlgorithm.h"
+#include "SlackNodeAlgorithm.h"
 
 #include <PARParametersSet.h>
 #include <boost/shared_ptr.hpp>
